+++ core/domain/interface/merchant/merchant.go
@@ -126,7 +126,7 @@ type (
 		CreateSignUpToken(memberId int32) string
 
 		// 根据商户申请密钥获取会员编号
-		GetMemberFromSignUpToken(token string) int64
+		GetMemberFromSignUpToken(token string) int32
 
 		// 提交商户注册信息
 		CommitSignUpInfo(*MchSignUp) (int32, error)
